refactor: スクリーンショット処理とOCR処理を分離し、日本語認識を改善
- scripts/screenshot.py: スクリーンショット取得処理を分離
- scripts/ocr.py: OCR処理を分離（日本語・英語対応）
- scripts/test_ocr.py: OCR処理テスト用スクリプト追加
- Vision Framework設定を改善:
  - 日本語(ja-JP)と英語(en-US)の両言語を認識
  - 高精度モード(VNRequestTextRecognitionLevelAccurate)を明示的に設定
  - 言語補正(UsesLanguageCorrection)を有効化
  - タイムアウトを30秒に延長（日本語認識のため）
- モジュール化により開発・テストが容易に
